+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * :sparkles: Collectionx adds asSequence method
 * :sparkles: Stringx adds asSequence method
 * :fire: MapBuilder rename build\* method
+* :fire: Rangex.until method now return \*Range